Move CLI binary to main.rs
