+++ src/main.rs
@@ -2,9 +2,46 @@ mod main_app;
 
 use main_app::App;
 use std::io;
+use structopt::StructOpt;
+use trad::handlers::echo_handler::EchoHandler;
+
+#[derive(Debug, Clone, StructOpt)]
+enum Command {
+    Echo {
+        #[structopt(short = "m", long = "message")]
+        message: String,
+    },
+}
+
+#[derive(Debug, Clone, StructOpt)]
+#[structopt(name = "Trad")]
+struct MainArgs {
+    #[structopt(subcommand)]
+    command: Option<Command>,
+
+    // -d or --debug, true if it is present
+    #[structopt(short = "d", long = "debug")]
+    debug: bool,
+}
 
 fn main() -> Result<(), io::Error> {
-    let mut app = App::new();
+    let args = MainArgs::from_args();
+    if args.debug {
+        println!("{:?}", args);
+    }
 
-    app.run()
+    if args.command.is_none() {
+        let mut app = App::new();
+
+        return app.run();
+    }
+
+    match args.command.unwrap() {
+        Command::Echo { message } => {
+            let handler = EchoHandler::new(message);
+            handler.run();
+
+            Ok(())
+        }
+    }
 }
